Minor cleanup of Monitor Mixin classes Moved column_pooler_mixin from support package to its logical place in monitor_mixin package
