+++ py/htm/advanced/algorithms/monitor_mixin/column_pooler_mixin.py
@@ -182,17 +182,17 @@ class ColumnPoolerMonitorMixin(MonitorMixinBase):
         """
         Returns plot of the cell activity.
     
-        @param title        (string)  an optional title for the figure
+        @param title (string) an optional title for the figure
     
-        @param showReset    (bool)    if true, the first set of cell activities
-                                      after a reset will have a gray background
+        @param showReset (bool) if true, the first set of cell activities
+                                after a reset will have a gray background
     
-        @param resetShading (float)   if showReset is true, this float specifies the
-                                      intensity of the reset background with 0.0
-                                      being white and 1.0 being black
+        @param resetShading (float) if showReset is true, this float specifies the
+                                    intensity of the reset background with 0.0
+                                    being white and 1.0 being black
     
-        @param activityType (string)  The type of cell activity to display. Valid
-                                      types include "activeCells"
+        @param activityType (string) The type of cell activity to display. Valid
+                                     types include "activeCells"
     
         @return (Plot) plot
         """
