+++ py/tests/advanced/algorithms/column_pooler_unit_test.py
@@ -20,7 +20,7 @@
 import unittest
 
 from htm.advanced.algorithms.column_pooler import ColumnPooler
-from htm.advanced.support.column_pooler_mixin import ColumnPoolerMonitorMixin
+from htm.advanced.algorithms.monitor_mixin.column_pooler_mixin import ColumnPoolerMonitorMixin
 
 
 class MonitoredColumnPooler(ColumnPoolerMonitorMixin, ColumnPooler):
